# simulations/booking/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Subway Booking Simulation",
    "description": "Simulated hotel booking site for Subway compiler development",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/reservations": {
      "post": {
        "summary": "Book Room",
        "description": "Create a reservation.",
        "operationId": "book_room_api_reservations_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BookingRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reservations/{res_id}": {
      "get": {
        "summary": "Get Reservation",
        "description": "Check reservation status.",
        "operationId": "get_reservation_api_reservations__res_id__get",
        "parameters": [
          {
            "name": "res_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Res Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/reservations/{res_id}/cancel": {
      "post": {
        "summary": "Cancel Reservation",
        "description": "Cancel a reservation and free up the dates.",
        "operationId": "cancel_reservation_api_reservations__res_id__cancel_post",
        "parameters": [
          {
            "name": "res_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Res Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReservationResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "BookingRequest": {
        "properties": {
          "room_id": {
            "type": "string",
            "title": "Room Id"
          },
          "check_in": {
            "type": "string",
            "title": "Check In"
          },
          "check_out": {
            "type": "string",
            "title": "Check Out"
          },
          "guest_name": {
            "type": "string",
            "title": "Guest Name"
          },
          "guest_email": {
            "type": "string",
            "title": "Guest Email"
          },
          "guest_phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Guest Phone"
          }
        },
        "type": "object",
        "required": [
          "room_id",
          "check_in",
          "check_out",
          "guest_name",
          "guest_email"
        ],
        "title": "BookingRequest"
      },
...
      "GuestInfo": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "title": "Email"
          },
          "phone": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Phone"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email"
        ],
        "title": "GuestInfo"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "Reservation": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "hotel_id": {
            "type": "string",
            "title": "Hotel Id"
          },
          "room_id": {
            "type": "string",
            "title": "Room Id"
          },
          "guest": {
            "$ref": "#/components/schemas/GuestInfo"
          },
          "check_in": {
            "type": "string",
            "title": "Check In"
          },
          "check_out": {
            "type": "string",
            "title": "Check Out"
          },
          "nights": {
            "type": "integer",
            "title": "Nights"
          },
          "total_price": {
            "type": "number",
            "title": "Total Price"
          },
          "status": {
            "$ref": "#/components/schemas/ReservationStatus",
            "default": "pending"
          },
          "created_at": {
            "type": "string",
            "title": "Created At",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "hotel_id",
          "room_id",
          "guest",
          "check_in",
          "check_out",
          "nights",
          "total_price"
        ],
        "title": "Reservation"
      },
      "ReservationResponse": {
        "properties": {
          "reservation": {
            "$ref": "#/components/schemas/Reservation"
          },
          "message": {
            "type": "string",
            "title": "Message",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "reservation"
        ],
        "title": "ReservationResponse"
      },
      "ReservationStatus": {
        "type": "string",
        "enum": [
          "pending",
          "confirmed",
          "cancelled"
        ],
        "title": "ReservationStatus"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}